refactor dpo trainer into own module
